pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "npu-webhook"
version = "0.1.0"
description = "个人知识库 + 记忆增强系统"
requires-python = ">=3.11"
license = "MIT"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "chromadb>=0.5",
    "onnxruntime>=1.19",
    "jieba>=0.42",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "watchdog>=5.0",
    "pymupdf>=1.24",
    "python-docx>=1.1",
    "markdown>=3.7",
    "apscheduler>=3.10",
    "jinja2>=3.1",
    "websockets>=13.0",
    "aiosqlite>=0.20",
    "tokenizers>=0.20",
    "python-multipart>=0.0.9",
]

[project.optional-dependencies]
npu-intel = ["openvino>=2024.0"]
npu-amd = ["onnxruntime-directml>=1.19"]
tray = [
    "pystray>=0.19",
    "Pillow>=10",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "ruff>=0.8",
    "mypy>=1.13",
]

[project.scripts]
npu-webhook = "npu_webhook.main:main"

[tool.hatch.build.targets.wheel]
packages = ["src/npu_webhook"]

[tool.ruff]
target-version = "py311"
line-length = 120
src = ["src"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
